Escritorio/AIAIAI_Consulting/projects/dev/moltbook-monitor/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Scraping (Playwright instead of Puppeteer - better Python support)
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# NLP & Analysis
afinn==0.1

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
